feat(SPV-1316): rename linking_key to a proper BRC number
