+++ engine/keys/type84/linking_key_test.go
@@ -1,4 +1,4 @@
-package type42
+package type84
 
 import (
 	"encoding/hex"
